Add files via upload
This file contains 3 files: first the code which is the updated one , containing everything related to the  analysis of the dataset, machine learning models tested etc. Then we have the story, done in tableau, contains useful graphs and charts related to the relations in the data set . Finally we have a comprehensive analysis , it is not yet done , still updating it , contains the analysis, interpretations etc.